requirements.txt
# ── Core deep learning ────────────────────────────────────────────────────────
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0

# ── Data & scientific ─────────────────────────────────────────────────────────
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0

# ── Visualisation ─────────────────────────────────────────────────────────────
matplotlib>=3.7.0
seaborn>=0.12.0
opencv-python-headless>=4.8.0

# ── Web application ───────────────────────────────────────────────────────────
gradio>=4.0.0

# ── Utilities ─────────────────────────────────────────────────────────────────
tqdm>=4.65.0
gdown>=4.7.0

# ── Optional: dataset download ───────────────────────────────────────────────
# kaggle>=1.5.0       # pip install kaggle  (requires ~/.kaggle/kaggle.json)

# ── Optional: ONNX export ────────────────────────────────────────────────────
# onnx>=1.14.0
# onnxscript>=0.1.0

# ── Development / testing ─────────────────────────────────────────────────────
pytest>=7.4.0
